packages/server/requirements.txt
fastapi
uvicorn
humps

sqlalchemy
PyMySQL

python-dotenv